Implement packaging put all classes to package and separate them according to the type
