+++ src/main/java/duke/Duke.java
@@ -1,15 +1,9 @@
-import java.time.DateTimeException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
-import java.io.File;
-import java.util.Scanner;
-import java.io.FileWriter;
-import java.io.IOException;
+package duke;
+
+import duke.command.Command;
+import duke.exception.CommandInvalidException;
+
 import java.io.FileNotFoundException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
 public class Duke {
     private Storage storage;
