+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,7 @@
-import java.io.IOException;
+package duke;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
