+++ src/main/java/duke/command/Commands.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public enum Commands {
     LIST, TODO, EVENT, DEADLINE, MARK, UNMARK, DELETE, BYE;
 }
