+++ src/main/java/duke/exception/CommandFormatException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class CommandFormatException extends Exception{
     public CommandFormatException(){
         super();
