+++ src/main/java/duke/exception/CommandInvalidException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class CommandInvalidException extends Exception{
     public CommandInvalidException(){
         super();
